backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
anthropic>=0.40.0
python-multipart==0.0.6
pydantic==2.5.3
python-dotenv==1.0.0
aiofiles==23.2.1
websockets==12.0
supabase==2.10.0
PyPDF2==3.0.1
